requirements.txt
flask
requests
python-dotenv
reportlab
flask_cors
pyinstaller
flask_socketio
eventlet
